refactor(tui): フィルタ modal を切り出して issue パッケージへまとめる
app.py が 1100 行を超え、issues タブ固有のレイアウトと TUI 全体の組み立てが
混在して見通しが悪くなっていた。

- フィルタ modal の Float 構築と描画を filter_modal.py へ分離し、app.py 側は
  build_filter_float() の呼び出し 1 行にする
- issues タブ関連のファイルを tui/issue/ へ移動し、テストも同じ構成に揃える

app.py は 1122 行から 1002 行になった。挙動の変更はない。
