+++ src/redi/tui/app.py
@@ -14,7 +14,6 @@ from prompt_toolkit.layout.containers import (
     Float,
     FloatContainer,
     HSplit,
-    ScrollOffsets,
     VSplit,
     Window,
 )
@@ -27,7 +26,8 @@ from redi.api.me import fetch_my_user_id
 from redi.api.membership import fetch_project_users
 from redi.api.project import fetch_projects
 from redi.i18n import messages
-from redi.tui.issue_tab import (
+from redi.tui.issue.filter_modal import build_filter_float
+from redi.tui.issue.issue_tab import (
     ISSUE_TAB,
     comment_select_cursor_down,
     comment_select_cursor_up,
@@ -41,8 +41,6 @@ from redi.tui.issue_tab import (
 )
 from redi.tui.state import (
     FIXED_ROWS,
-    FilterField,
-    FilterModalState,
     IssueFilter,
     Renderable,
     TimeEntryFilter,
@@ -292,90 +290,6 @@ def _render_project_modal(state: TuiState) -> Renderable:
         parts.append((line_style, f" {cursor_mark} {active_mark} {label}\n"))
     parts.append(("", messages.tui_project_modal_hint))
     return parts
-
-
-def _render_filter_section(
-    modal: FilterModalState,
-    section: FilterField,
-    title: str,
-    choices: list[tuple[str | None, str]],
-    cursor: int,
-    active_id: str | None,
-) -> Renderable:
-    focused = modal.focus == section
-    header_style = "bold fg:ansicyan" if focused else "bold"
-    parts: Renderable = [(header_style, f"[{title}]\n")]
-    for i, (api_val, label) in enumerate(choices):
-        is_cursor = focused and i == cursor
-        is_active = api_val == active_id
-        cursor_mark = ">" if is_cursor else " "
-        active_mark = "*" if is_active else " "
-        line_style = "reverse" if is_cursor else ("bold" if is_active else "")
-        parts.append((line_style, f" {cursor_mark} {active_mark} {label}\n"))
-    return parts
-
-
-def _render_filter_column(state: TuiState, section: FilterField) -> Renderable:
-    """フィルタ modal の 1 列 (status か assignee) を描画する。
-
-    2 列を縦に連結せず列ごとに描くことで、modal の高さが選択肢数の合計ではなく
-    max(status, assignee) で済み、選択肢が多くても縦に溢れにくくなる。
-    """
-    f = state.issue_tab.filter
-    modal = state.issue_tab.filter_modal
-    if section == "status":
-        return _render_filter_section(
-            modal,
-            "status",
-            messages.tui_filter_status,
-            modal.status_choices,
-            modal.status_cursor,
-            f.status_id,
-        )
-    return _render_filter_section(
-        modal,
-        "assignee",
-        messages.tui_filter_assignee,
-        modal.assignee_choices,
-        modal.assignee_cursor,
-        f.assigned_to_id,
-    )
-
-
-def _filter_column_cursor_y(modal: FilterModalState, section: FilterField) -> int:
-    """`_render_filter_column` の描画結果におけるカーソル行 (0 始まり)。
-
-    Window にカーソル位置を伝えて選択中の行が常に画面内へ来るようスクロール
-    させるために使う。0 行目はセクションヘッダなので選択肢は 1 行目から並ぶ。
-    """
-    cursor = modal.status_cursor if section == "status" else modal.assignee_cursor
-    return 1 + cursor
-
-
-def _render_filter_hint() -> Renderable:
-    return [("", messages.tui_filter_hint)]
-
-
-def _filter_column_window(state: TuiState, section: FilterField) -> Window:
-    """フィルタ modal の 1 列を載せる Window。
-
-    選択肢が端末高を超えると Float が高さを端末内へ切り詰め、Window にはその
-    切り詰め後の高さが渡る。`get_cursor_position` を与えておくと Window が
-    カーソル行を画面内に収めるようスクロールしてくれるので、選択肢が多くても
-    選択中の行を見失わない (渡さないと vertical_scroll が 0 のまま先頭が出続け、
-    下の方の選択肢が見えなくなる)。
-    """
-    return Window(
-        FormattedTextControl(
-            lambda: _render_filter_column(state, section),
-            show_cursor=False,
-            get_cursor_position=lambda: Point(
-                0, _filter_column_cursor_y(state.issue_tab.filter_modal, section)
-            ),
-        ),
-        wrap_lines=False,
-        scroll_offsets=ScrollOffsets(top=1, bottom=1),
-    )
 
 
 def _render_time_entry_filter_modal(state: TuiState) -> Renderable:
@@ -1042,40 +956,7 @@ def run_issue_tui(
         ),
     )
 
-    filter_float = Float(
-        content=ConditionalContainer(
-            content=VSplit(
-                [
-                    Window(width=1, char=" "),
-                    Frame(
-                        HSplit(
-                            [
-                                VSplit(
-                                    [
-                                        _filter_column_window(state, "status"),
-                                        Window(width=1, char=" "),
-                                        Window(width=1, char="│"),
-                                        Window(width=1, char=" "),
-                                        _filter_column_window(state, "assignee"),
-                                    ]
-                                ),
-                                # ヒントは列のスクロール対象から外して常に見せる
-                                Window(
-                                    FormattedTextControl(
-                                        _render_filter_hint, show_cursor=False
-                                    ),
-                                    height=1,
-                                ),
-                            ]
-                        ),
-                        title=messages.tui_filter_title,
-                    ),
-                    Window(width=1, char=" "),
-                ]
-            ),
-            filter=show_filter_modal,
-        ),
-    )
+    filter_float = build_filter_float(state, show_filter_modal)
 
     time_entry_filter_float = Float(
         content=ConditionalContainer(
