+++ src/redi/tui/issue/issue_tab.py
