+++ tests/unit/tui/test_project_switch.py
@@ -9,7 +9,8 @@ from redi import config
 from redi.api.project import Project
 from redi.api.time_entry import TimeEntry
 from redi.i18n import messages
-from redi.tui import app, issue_tab, time_entry_tab
+from redi.tui import app, time_entry_tab
+from redi.tui.issue import issue_tab
 from redi.tui.state import IssueFilter, TimeEntryFilter, TuiState
 from redi.tui.tab import TabView, noop, noop_jump
