+++ tests/unit/tui/test_tab_reload.py
@@ -5,7 +5,8 @@ from typing import cast
 from redi.api.issue import Issue
 from redi.api.time_entry import TimeEntry
 from redi.api.wiki import WikiPage
-from redi.tui import issue_tab, time_entry_tab, wiki_tab
+from redi.tui import time_entry_tab, wiki_tab
+from redi.tui.issue import issue_tab
 from redi.tui.state import TuiState
